docs: extract function to util/dom
